# Optimize memcpy, memmove and memset