Partially addresses #339. (memcmp is not implemented in this PR)

This PR modifies the current simple implementation of memcpy, memmove and memset with a more sophisticated one.
It will first align dest to machine-word boundary using bytewise copy.
If dest and src are co-aligned (i.e. dest and src address are congruent modular machine-word size), it will proceed to use machine-word-wide copy to copy as many bytes as possible. Remaining bytes are copied using bytewise copy.
If dest and src are not co-aligned, misaligned copying is performed by reading two adjacent machine words and assemble them together with logical shifts before writing to dest.

The existing implementation can be pretty fast on platforms with SIMD and vectorization, but falls short for those without. The code in this PR is carefully written so it is fast on those platforms while is still vectorizable.

Here are some performance numbers:

<details>
<summary>Rust Builtin</summary>

```
test memcpy_builtin_1048576        ... bench:      33,388 ns/iter (+/- 2,619) = 31405 MB/s
test memcpy_builtin_1048576_offset ... bench:      33,988 ns/iter (+/- 2,371) = 30851 MB/s
test memcpy_builtin_4096           ... bench:          37 ns/iter (+/- 2) = 110702 MB/s
test memcpy_builtin_4096_offset    ... bench:          37 ns/iter (+/- 1) = 110702 MB/s
test memmove_builtin_1048576       ... bench:      28,862 ns/iter (+/- 1,013) = 36330 MB/s
test memmove_builtin_4096          ... bench:          38 ns/iter (+/- 1) = 107789 MB/s
test memset_builtin_1048576        ... bench:      21,171 ns/iter (+/- 1,369) = 49528 MB/s
test memset_builtin_1048576_offset ... bench:      21,516 ns/iter (+/- 906) = 48734 MB/s
test memset_builtin_4096           ... bench:          62 ns/iter (+/- 0) = 66064 MB/s
test memset_builtin_4096_offset    ... bench:          68 ns/iter (+/- 2) = 60235 MB/s
```

</details>

<details>
<summary>New Implementation</summary>

```
test memcpy_rust_1048576           ... bench:      40,900 ns/iter (+/- 2,602) = 25637 MB/s
test memcpy_rust_1048576_offset    ... bench:      46,575 ns/iter (+/- 2,994) = 22513 MB/s
test memcpy_rust_4096              ... bench:          76 ns/iter (+/- 4) = 53894 MB/s
test memcpy_rust_4096_offset       ... bench:          95 ns/iter (+/- 3) = 43115 MB/s
test memmove_rust_1048576          ... bench:      50,363 ns/iter (+/- 2,829) = 20820 MB/s
test memmove_rust_4096             ... bench:         176 ns/iter (+/- 6) = 23272 MB/s
test memset_rust_1048576           ... bench:      27,171 ns/iter (+/- 1,168) = 38591 MB/s
test memset_rust_1048576_offset    ... bench:      36,623 ns/iter (+/- 1,418) = 28631 MB/s
test memset_rust_4096              ... bench:          74 ns/iter (+/- 5) = 55351 MB/s
test memset_rust_4096_offset       ... bench:          95 ns/iter (+/- 2) = 43115 MB/s
```

</details>

<details>
<summary>Old Implementation</summary>

```
test memcpy_rust_1048576           ... bench:      43,626 ns/iter (+/- 5,963) = 24035 MB/s
test memcpy_rust_1048576_offset    ... bench:      47,138 ns/iter (+/- 4,029) = 22244 MB/s
test memcpy_rust_4096              ... bench:          73 ns/iter (+/- 3) = 56109 MB/s
test memcpy_rust_4096_offset       ... bench:          90 ns/iter (+/- 3) = 45511 MB/s
test memmove_rust_1048576          ... bench:     305,743 ns/iter (+/- 12,242) = 3429 MB/s
test memmove_rust_4096             ... bench:       1,071 ns/iter (+/- 60) = 3824 MB/s
test memset_rust_1048576           ... bench:      28,485 ns/iter (+/- 2,085) = 36811 MB/s
test memset_rust_1048576_offset    ... bench:      36,999 ns/iter (+/- 316) = 28340 MB/s
test memset_rust_4096              ... bench:          69 ns/iter (+/- 0) = 59362 MB/s
test memset_rust_4096_offset       ... bench:          90 ns/iter (+/- 9) = 45511 MB/s
```

</details>

<details>
<summary>New, with no-vectorize-loops</summary>

```
test memcpy_rust_1048576           ... bench:      70,713 ns/iter (+/- 24,368) = 14828 MB/s
test memcpy_rust_1048576_offset    ... bench:      68,340 ns/iter (+/- 5,622) = 15343 MB/s
test memcpy_rust_4096              ... bench:         272 ns/iter (+/- 15) = 15058 MB/s
test memcpy_rust_4096_offset       ... bench:         279 ns/iter (+/- 7) = 14681 MB/s
test memmove_rust_1048576          ... bench:      68,726 ns/iter (+/- 3,034) = 15257 MB/s
test memmove_rust_4096             ... bench:         275 ns/iter (+/- 10) = 14894 MB/s
test memset_rust_1048576           ... bench:      42,382 ns/iter (+/- 1,542) = 24741 MB/s
test memset_rust_1048576_offset    ... bench:      42,585 ns/iter (+/- 1,722) = 24623 MB/s
test memset_rust_4096              ... bench:         145 ns/iter (+/- 5) = 28248 MB/s
test memset_rust_4096_offset       ... bench:         146 ns/iter (+/- 7) = 28054 MB/s
```

</details>

<details>
<summary>Old, with no-vectorize-loops</summary>

```
test memcpy_rust_1048576           ... bench:     306,552 ns/iter (+/- 19,395) = 3420 MB/s
test memcpy_rust_1048576_offset    ... bench:     311,564 ns/iter (+/- 14,247) = 3365 MB/s
test memcpy_rust_4096              ... bench:       1,076 ns/iter (+/- 63) = 3806 MB/s
test memcpy_rust_4096_offset       ... bench:       1,074 ns/iter (+/- 30) = 3813 MB/s
test memmove_rust_1048576          ... bench:     315,208 ns/iter (+/- 158,448) = 3326 MB/s
test memmove_rust_4096             ... bench:       1,077 ns/iter (+/- 33) = 3803 MB/s
test memset_rust_1048576           ... bench:     303,456 ns/iter (+/- 7,937) = 3455 MB/s
test memset_rust_1048576_offset    ... bench:     304,886 ns/iter (+/- 21,495) = 3439 MB/s
test memset_rust_4096              ... bench:       1,070 ns/iter (+/- 30) = 3828 MB/s
test memset_rust_4096_offset       ... bench:       1,064 ns/iter (+/- 99) = 3849 MB/s
```

</details>

All above numbers are tested on my Coffee Lake laptop, with no-asm feature on.

The number shows that if vectorization is enabled or supported on the platform, then memcpy and memset performance of the old and the new implementation are similar (while much faster on memmove, it seems the compiler had a hard time vectorizing backward copy). If there is no vectorization or the platform does not supported misaligned access, the new implementation will have very significant performance gain (>4x with the number above, and on in-order scaler cores you'd expect 3~8x improvement depending on whether access is aligned and the width of usize).

I inspected the code generated assembly for RISC-V and it seems pretty close in quality to my hand optimsied assembly.