@nbdd0121 2021-02-10
Misaligned benchmarks:

<details>
<summary>Rust Builtin</summary>

```
test memcpy_builtin_1048576_misalign  ... bench:      33,722 ns/iter (+/- 1,310) = 31094 MB/s
test memcpy_builtin_4096_misalign     ... bench:          38 ns/iter (+/- 1) = 107789 MB/s
test memmove_builtin_1048576_misalign ... bench:      34,791 ns/iter (+/- 1,177) = 30139 MB/s
test memmove_builtin_4096_misalign    ... bench:          38 ns/iter (+/- 1) = 107789 MB/s
```

</details>

<details>
<summary>New Implementation</summary>

```
test memcpy_rust_1048576_misalign     ... bench:      56,567 ns/iter (+/- 1,812) = 18536 MB/s
test memcpy_rust_4096_misalign        ... bench:         215 ns/iter (+/- 5) = 19051 MB/s
test memmove_rust_1048576_misalign    ... bench:     136,350 ns/iter (+/- 5,338) = 7690 MB/s
test memmove_rust_4096_misalign       ... bench:         550 ns/iter (+/- 27) = 7447 MB/s
```

</details>

<details>
<summary>Old Implementation</summary>

```
test memcpy_rust_1048576_misalign     ... bench:      49,171 ns/iter (+/- 3,514) = 21325 MB/s
test memcpy_rust_4096_misalign        ... bench:         108 ns/iter (+/- 5) = 37925 MB/s
test memmove_rust_1048576_misalign    ... bench:     306,153 ns/iter (+/- 10,100) = 3425 MB/s
test memmove_rust_4096_misalign       ... bench:       1,070 ns/iter (+/- 31) = 3828 MB/s
```

</details>
<details>
<summary>New, with no-vectorize-loops</summary>

```
test memcpy_rust_1048576_misalign     ... bench:     138,407 ns/iter (+/- 5,799) = 7576 MB/s
test memcpy_rust_4096_misalign        ... bench:         544 ns/iter (+/- 41) = 7529 MB/s
test memmove_rust_1048576_misalign    ... bench:     139,043 ns/iter (+/- 5,207) = 7541 MB/s
test memmove_rust_4096_misalign       ... bench:         551 ns/iter (+/- 8) = 7433 MB/s
```

</details>
<details>
<summary>Old, with no-vectorize-loops</summary>

```
test memcpy_rust_1048576_misalign     ... bench:     308,055 ns/iter (+/- 5,330) = 3403 MB/s
test memcpy_rust_4096_misalign        ... bench:       1,073 ns/iter (+/- 13) = 3817 MB/s
test memmove_rust_1048576_misalign    ... bench:     308,875 ns/iter (+/- 27,410) = 3394 MB/s
test memmove_rust_4096_misalign       ... bench:       1,074 ns/iter (+/- 36) = 3813 MB/s
```

</details>

I expected the misaligned code path hurts x86-64. It is slower with vectorization on but faster with vectorization off. It is always faster in memmove, presumably because LLVM couldn't vectorize it.