@nbdd0121 2021-02-17
I would argue this isn't overly complicated. The concept is very simple and is what C libraries do already.

Benchmarking on ARM/AArch64 might not be very helpful because there are both vectorization and misaligned load/store support, but I would still expect memmove performance improvement (I could test it myself because I don't have a ARM dev environment setup). This PR is mainly intended to be a much better default for those without (a quick test on godbolt suggests this includes armv7, powerpc, mips, riscv and more).

As for the CI, it should be noted a few platforms are already being tested (including big-endian ones like PowerPC).