@adamgemmell 2021-08-20
Hi, I've carried out some benchmarks on an aarch64-unknown-linux-gnu platform (Graviton 2, Neoverse-N1) to alleviate some concerns. I rebased the branch to something more recent before carrying these out.

(all numbers ns/iter. +/- ranges were never more than 1% of the result)

Test Name | Result
memcpy_builtin_104857 | 45562
memcpy_builtin_1048576_misalign | 47097
memcpy_builtin_1048576_offset | 47052
memcpy_builtin_4096 | 108
memcpy_builtin_4096_misalign | 111
memcpy_builtin_4096_offset | 111
memmove_builtin_1048576 | 42840
memmove_builtin_1048576_misalign | 62671
memmove_builtin_4096 | 109
memmove_builtin_4096_misalign | 161
memset_builtin_1048576 | 26245
memset_builtin_1048576_offset | 26242
memset_builtin_4096 | 105
memset_builtin_4096_offset | 106

Test Name | New | New (no-vec) | Old | Old (no-vec)
-- | -- | -- | -- | --
memcpy_rust_1048576 | 47317 | 108806 | 46978 | 709112
memcpy_rust_1048576_misalign | 78310 | 148440 | 56356 | 708612
memcpy_rust_1048576_offset | 50758 | 109122 | 57212 | 708791
memcpy_rust_4096 | 113 | 421 | 109 | 2766
memcpy_rust_4096_misalign | 287 | 573 | 136 | 2768
memcpy_rust_4096_offset | 121 | 426 | 136 | 2768
memmove_rust_1048576 | 105688 | 109360 | 840238 | 630534
memmove_rust_1048576_misalign | 151630 | 151128 | 840360 | 630725
memmove_rust_4096 | 392 | 421 | 3288 | 2469
memmove_rust_4096_misalign | 573 | 568 | 3288 | 2469
memset_rust_1048576 | 26252 | 104967 | 26246 | 419879
memset_rust_1048576_offset | 26257 | 104960 | 26247 | 419857
memset_rust_4096 | 109 | 420 | 106 | 1658
memset_rust_4096_offset | 114 | 422 | 106 | 1654


The short story is the results look similar to the ones observed above on x86_64.

- memcpy and memset are largely the same, with nice improvements for non-vectorised code. Performance reaches that of the builtins in at least some cases.
- memmove shows large improvments. Non-vectorised performance seems comparable at least for these tests.
- Misaligned accesses usually indicate a performance regression, though it's smaller than the improvement in memmove.